% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_mortgage_rates}
\alias{get_mortgage_rates}
\title{Retrieve Mortgage Rates from FRED}
\usage{
get_mortgage_rates()
}
\value{
A table of class \code{data.frame} that contains monthly mortgage rate
data.
}
\description{
This function is designed to return monthly mortgage rates provided by the
St. Louis Federal Reserve's Economic Data (FRED) system. Specifically,
this function will return a \code{data.frame} containing the "MORTG" FRED
data series, which is the monthly spot rates for 30-year mortgage loans in
the United States.
}
\details{
This function is not designed to accept any arguments; you may wish to
inspect the documentation of the \code{\link[quantmod]{getSymbols}} function
from the \code{quantmod} package for more flexible financial & economic data
retrieval.
}
