% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{zillow_get_url}
\alias{zillow_get_url}
\title{Determine static web URL address for Zillow data}
\usage{
zillow_get_url(zillow_var, geo_level)
}
\arguments{
\item{zillow_var}{The variable name to read in. The full list of these
short-form variable names can be found in the variable dictionary included
with this package (i.e. by calling \code{zillow_var_dict})}

\item{geo_level}{The geographic level of the data to retrieve, passed as a
character string. Available options are: \code{state}, \code{metro}, \code{county},
\code{city}, \code{zip}, and \code{neighborhood}.}
}
\description{
This function is designed to fetch the full URL of Zillow's public research
data via the site's static-address \code{.csv} links. It is specificaly designed
as a utility in actually fetching the data.
}
