% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{zillow_read}
\alias{zillow_read}
\title{Read static Zillow data from the web}
\usage{
zillow_read(zillow_var, geo_level, download_dir = "~/Downloads")
}
\arguments{
\item{zillow_var}{The variable name to read in. The full list of these
short-form variable names can be found in the variable dictionary included
with this package (i.e. by calling \code{zillow_var_dict})}

\item{geo_level}{The geographic level of the data to retrieve, passed as a
character string. Available options are: \code{state}, \code{metro}, \code{county},
\code{city}, \code{zip}, and \code{neighborhood}.}

\item{download_dir}{Directory to save Zillow data to. Defaults to attempt the
standard path of \code{~/Downloads}.}
}
\value{
A table of class \code{data.frame}
}
\description{
This function is designed to read in Zillow's public research data via the
site's static-address \code{.csv} links.
}
