% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{zillow_read_from_web}
\alias{zillow_read_from_web}
\title{Read static Zillow data from the web}
\usage{
zillow_read_from_web(zillow_var, geo_level)
}
\arguments{
\item{zillow_var}{The variable name to read in. The full list of
these short-form variable names can be found in the \code{zillow_var_ids}
data available in this package.}
}
\value{
A table of class \code{tbl_df}
}
\description{
This function is designed to read in Zillow's public research
data via the site's static-address \code{.csv} links.
}
